Revert "Merge pull request #215 from alesj/jmx_int"
This reverts commit e9a18d76421d0bc70bb47d3a18cb7fd70baec70b, reversing
changes made to e2751123b7cc9d81d7d58a7e47d7768260186e40.
