--- jmx/src/main/java/org/jboss/as/jmx/DelegateMBeanServerPlugin.java
@@ -1,53 +1,0 @@
-/*
- * JBoss, Home of Professional Open Source.
- * Copyright 2014, Red Hat, Inc., and individual contributors
- * as indicated by the @author tags. See the copyright.txt file in the
- * distribution for a full listing of individual contributors.
- *
- * This is free software; you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation; either version 2.1 of
- * the License, or (at your option) any later version.
- *
- * This software is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this software; if not, write to the Free
- * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
- * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
- */
-
-package org.jboss.as.jmx;
-
-import javax.management.ObjectName;
-
-import org.jboss.as.server.jmx.MBeanServerPlugin;
-
-/**
- * @author Ales Justin
- */
-public class DelegateMBeanServerPlugin extends DelegateMBeanServer implements MBeanServerPlugin {
-
-    public DelegateMBeanServerPlugin(MBeanServerPlugin delegate) {
-        super(delegate);
-    }
-
-    protected MBeanServerPlugin getDelegate() {
-        return (MBeanServerPlugin) super.getDelegate();
-    }
-
-    public boolean accepts(ObjectName objectName) {
-        return getDelegate().accepts(objectName);
-    }
-
-    public boolean shouldAuditLog() {
-        return getDelegate().shouldAuditLog();
-    }
-
-    public boolean shouldAuthorize() {
-        return getDelegate().shouldAuthorize();
-    }
-}
